+++ docs/en/changes/changes.md
@@ -93,6 +93,7 @@
 * Remove method `appendMutant` from StorageID.
 * Fix otlp log handler reponse error and otlp span convert error.
 * Fix service_relation source layer in mq entry span analyse.
+* Fix metrics comparison in promql with bool modifier.
 
 #### UI